python/google-adk/cloud-run-invoicing-agent/requirements.txt
google-adk==1.2.1
google-cloud-aiplatform==1.96.0
google-genai==1.19.0
fastapi>=0.115.0
uvicorn==0.34.3
python-dotenv==1.0.1
